+++ plugins/tailwindcss/Contracts/Plugin.d.ts
@@ -1,0 +1,4 @@
+export interface PluginContract
+{
+    build(): void;
+}