+++ plugins/tailwindcss/index.ts
@@ -1,12 +1,11 @@
-import type { PluggerOptions } from "../../types";
 import Plugger from "./Plugger";
 import plugin from "tailwindcss/plugin";
 import type { PluginAPI } from "tailwindcss/types/config";
+import { UIOptions } from "./types";
 
-export const UIPlugin = plugin.withOptions((options: PluggerOptions = {}) =>
-    (api: PluginAPI) => {
-        const plugger = new Plugger(api, options);
-        plugger.plug();
-    });
+export const UIPlugin = plugin.withOptions((options: UIOptions = undefined) => (api: PluginAPI) => {
+    const plugger = new Plugger(api);
+    plugger.boot(options);
+});
 
 export default UIPlugin;